[Tests] Move tiny model construction to tests/helpers/tiny_model.py
Make build_tiny_from_configs a reusable utility in tests/helpers/ so
other tests can use it. diff_model_builders.py now just contains the
per-model one-liner builders.
